@@ -11,71 +11,70 @@
   "homepage": "http://ebenmonney.com",
   "repository": "https://github.com/emonney/quickapp.git",
   "dependencies": {
-    "@angular/animations": "4.3.6",
-    "@angular/common": "4.3.6",
-    "@angular/compiler": "4.3.6",
-    "@angular/compiler-cli": "4.3.6",
-    "@angular/core": "4.3.6",
-    "@angular/forms": "4.3.6",
-    "@angular/http": "4.3.6",
-    "@angular/platform-browser": "4.3.6",
-    "@angular/platform-browser-dynamic": "4.3.6",
-    "@angular/platform-server": "4.3.6",
-    "@angular/router": "4.3.6",
-    "@ngtools/webpack": "1.7.1",
-    "@types/webpack-env": "1.13.1",
-    "@ngx-translate/core": "8.0.0",
-    "@swimlane/ngx-datatable": "10.2.3",
-    "@types/jquery": "3.2.12",
+    "@angular/animations": "5.1.0",
+    "@angular/common": "5.1.0",
+    "@angular/compiler": "5.1.0",
+    "@angular/compiler-cli": "5.1.0",
+    "@angular/core": "5.1.0",
+    "@angular/forms": "5.1.0",
+    "@angular/platform-browser": "5.1.0",
+    "@angular/platform-browser-dynamic": "5.1.0",
+    "@angular/platform-server": "5.1.0",
+    "@angular/router": "5.1.0",
+    "@ngtools/webpack": "1.9.0",
+    "@types/webpack-env": "1.13.3",
+    "@ngx-translate/core": "9.0.1",
+    "@swimlane/ngx-datatable": "11.1.5",
+    "@types/jquery": "3.2.16",
     "angular2-template-loader": "0.6.2",
     "aspnet-webpack": "2.0.1",
-    "awesome-typescript-loader": "3.2.3",
+    "awesome-typescript-loader": "3.4.1",
     "bootstrap": "3.3.7",
     "bootstrap-datepicker": "1.7.1",
     "bootstrap-select": "1.12.4",
     "bootstrap-toggle": "2.2.2",
     "bootstrap-vertical-tabs": "1.2.2",
-    "chart.js": "2.7.0",
-    "core-js": "2.5.1",
+    "chart.js": "2.7.1",
+    "core-js": "2.5.3",
     "css": "2.2.1",
     "css-loader": "0.28.7",
-    "event-source-polyfill": "0.0.9",
-    "expose-loader": "0.7.3",
-    "extract-text-webpack-plugin": "3.0.0",
-    "file-loader": "0.11.2",
+    "event-source-polyfill": "0.0.12",
+    "expose-loader": "0.7.4",
+    "extract-text-webpack-plugin": "3.0.2",
+    "file-loader": "1.1.5",
     "font-awesome": "4.7.0",
     "html-loader": "0.5.1",
     "jquery": "3.2.1",
     "json-loader": "0.5.7",
     "ng2-charts": "1.6.0",
     "ng2-toasty": "4.0.3",
-    "ngx-bootstrap": "1.9.3",
-    "node-sass": "4.5.3",
+    "ngx-bootstrap": "2.0.0-beta.11",
+    "node-sass": "4.7.2",
     "raw-loader": "0.5.1",
-    "rxjs": "5.4.3",
+    "rxjs": "5.5.5",
     "sass-loader": "6.0.6",
-    "style-loader": "0.18.2",
+    "style-loader": "0.19.0",
     "to-string-loader": "1.1.5",
-    "typescript": "2.5.2",
-    "url-loader": "0.5.9",
+    "typescript": "2.6.2",
+    "url-loader": "0.6.2",
     "web-animations-js": "2.3.1",
-    "webpack": "3.6.0",
-    "webpack-hot-middleware": "2.19.1",
-    "webpack-merge": "4.1.0",
-    "zone.js": "0.8.17"
+    "webpack": "3.10.0",
+    "webpack-hot-middleware": "2.21.0",
+    "webpack-merge": "4.1.1",
+    "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@types/chai": "4.0.4",
-    "@types/jasmine": "2.6.0",
+    "@types/chai": "4.0.10",
+    "@types/jasmine": "2.8.2",
     "chai": "4.1.2",
     "jasmine-core": "2.8.0",
     "karma": "1.7.1",
     "karma-chai": "0.1.0",
     "karma-chrome-launcher": "2.2.0",
     "karma-cli": "1.0.1",
-    "karma-jasmine": "1.1.0",
+    "karma-jasmine": "1.1.1",
     "karma-jasmine-html-reporter": "0.2.2",
-    "karma-webpack": "2.0.4"
+    "karma-webpack": "2.0.6"
   },
   "scripts": {
     "dev-build": "node node_modules/webpack/bin/webpack.js --config webpack.config.vendor.js",
