@@ -30,8 +30,8 @@ export class ConfigurationService {
 
     public static readonly appVersion: string = "1.0.0";
 
-    public baseUrl: string = Utilities.baseUrl().replace(/\/$/, '');
-    public fallbackBaseUrl: string = "http://ebenmonney.com/demo/quickapp";
+    public baseUrl: string = Utilities.baseUrl();
+    public fallbackBaseUrl: string = "http://quickapp.ebenmonney.com";
     public loginUrl: string = "/Login";
 
     //***Specify default configurations here***
